runtime fix

+++ code/game/objects/structures/burrows.dm
@@ -62,7 +62,7 @@
 	.=..()
 	all_burrows.Add(src)
 	var/obj/machinery/power/nt_obelisk/obelisk = locate(/obj/machinery/power/nt_obelisk) in range(7, src)
-	if(obelisk || obelisk.active)
+	if(obelisk && obelisk.active)
 		qdel(src)
 		return
 	if (anchor)
